fix small issues, begin to implement timeout perturbator

+++ perturbation_agent/src/main/java/se/kth/chaos/pagent/examples/ArrayPOneOnTTorrent.java
@@ -12,7 +12,7 @@ public class ArrayPOneOnTTorrent {
         String javaagentPath = System.getProperty("user.dir") + "/target/foagent-perturbation-jar-with-dependencies.jar";
         String endingPattern = "BitTorrent client signing off";
         String threadName = "ttorrent-1.5-client.jar";
-        String targetCsv = "perturbationPointsList.csv";
+        String targetCsv = "perturbationPointsList_ref.csv";
         String correctChecksum = "812ac191b8898b33aed4aef9ab066b5a";
         String osName = System.getProperty("os.name");
 
@@ -23,6 +23,7 @@ public class ArrayPOneOnTTorrent {
             File targetFile = null;
             for (int i = 1; i < perturbationInfo.size(); i++) {
                 String[] info = perturbationInfo.get(i);
+                if (!info[10].equals("0")) continue;
 
                 targetFile = new File(rootPath + "/ubuntu-14.04.5-server-i386.iso");
                 if (targetFile.exists()) { targetFile.delete(); }
@@ -33,7 +34,8 @@ public class ArrayPOneOnTTorrent {
                 System.out.println("start to perturb with array_pone at " + location);
 
                 try {
-                    String command = String.format("timeout --signal=9 300 java -javaagent:%s=mode:array_pone,filter:%s -jar %s -o . -s 0 ubuntu-14.04.5-server-i386.iso.torrent 2>&1", javaagentPath, location.replace("$", "\\$"), threadName);
+                    String command = String.format("timeout --signal=9 300 java -javaagent:%s=mode:array_pone,defaultMode:array_pone,rate:%s,countdown:1,filter:%s -jar %s -o . -s 0 ubuntu-14.04.5-server-i386.iso.torrent 2>&1", javaagentPath, info[6], location.replace("$", "\\$"), threadName);
+                    // System.out.println(command);
                     process = Runtime.getRuntime().exec(new String[]{"bash", "-c", command}, null, new File(rootPath));
 
                     InputStream inputStream = process.getInputStream();
@@ -43,7 +45,7 @@ public class ArrayPOneOnTTorrent {
                     int injectionCount = 0;
 
                     while((line = bufferedReader.readLine()) != null) {
-                        if (line.startsWith("INFO PAgent add an array_pone perturbator into")) {
+                        if (line.startsWith("INFO PAgent array_pone perturbation activated in")) {
                             injectionCount++;
                         } else if (line.contains(endingPattern)) {
                             process.destroy();
